talkie_bot/requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.26.0
drf-spectacular-sidecar==2023.3.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
mysqlclient==2.1.1
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
